[skip ci] extract client code to separate module
